+++ src/test/kotlin/onestep/racingcar/domain/RacingTest.kt
@@ -1,4 +1,4 @@
-package onestep.racingcar.model
+package onestep.racingcar.domain
 
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
